в связи с изменением типа id были переписаны функции

+++ src/main/java/by/sapra/listofcontacts/services/impl/JdbcContactService.java
@@ -55,7 +55,7 @@ class JdbcContactService implements ContactService {
 
     private ContactEntity mapToEntity(ContactPayload payload) {
         return ContactEntity.builder()
-                .id(payload.getId() != null ? Integer.valueOf(payload.getId()) : null)
+                .id(payload.getId() != null ? Long.valueOf(payload.getId()) : null)
                 .phone(payload.getPhone())
                 .lastName(payload.getLastName())
                 .firstName(payload.getFirstName())
